Moved constants.h and utils.h to the include/
