+++ include/rnp.h
@@ -33,7 +33,7 @@
 
 #include <stddef.h>
 #include "packet.h"
-#include "../common/constants.h"
+#include "constants.h"
 
 #ifndef __BEGIN_DECLS
 #if defined(__cplusplus)
