+++ src/lib/create.c
@@ -95,7 +95,7 @@ __RCSID("$NetBSD: create.c,v 1.38 2010/11/15 08:03:39 agc Exp $");
 #include "packet-key.h"
 #include "rnpsdk.h"
 #include "ecdsa.h"
-#include "../common/constants.h"
+#include "constants.h"
 
 extern ec_curve_desc_t ec_curves[PGP_CURVE_MAX];
