+++ src/lib/eddsa.c
@@ -28,8 +28,8 @@
 #include "bn.h"
 #include <string.h>
 #include <botan/ffi.h>
-#include "../common/utils.h"
-#include "../common/constants.h"
+#include "utils.h"
+#include "constants.h"
 
 int
 pgp_genkey_eddsa(pgp_seckey_t *seckey, size_t curve_len)
