+++ src/lib/packet-parse.c
@@ -92,7 +92,7 @@ __RCSID("$NetBSD: packet-parse.c,v 1.51 2012/03/05 02:20:18 christos Exp $");
 #include "crypto.h"
 #include "rnpdigest.h"
 #include "s2k.h"
-#include "../common/utils.h"
+#include "utils.h"
 
 #define ERRP(cbinfo, cont, err)                    \
     do {                                           \
